backend/enrichers/requirements.txt
azure-functions
pyodbc
azure-identity
azure-storage-blob

httpx
azure-servicebus
python-dotenv